+++ src/main/java/com/kr1s1s/subtlyd/client/model/geom/ModelLayersSD.java
@@ -28,6 +28,8 @@ public class ModelLayersSD extends ModelLayers {
     public static final ModelLayerLocation RED_TENT = register("tent/red_tent");
     public static final ModelLayerLocation BLACK_TENT = register("tent/black_tent");
     public static final ModelLayerLocation UNLIT_CAMPFIRE = register("unlit_campfire");
+    public static final ModelLayerLocation POTTAGE = register("pottage");
+
 
     public static List<ModelLayerLocation> TENTS = List.of(WHITE_TENT, ORANGE_TENT, MAGENTA_TENT, LIGHT_BLUE_TENT, YELLOW_TENT, LIME_TENT, PINK_TENT, GRAY_TENT, LIGHT_GRAY_TENT, CYAN_TENT, PURPLE_TENT, BLUE_TENT, BROWN_TENT, GREEN_TENT, RED_TENT, BLACK_TENT);